requirements.txt
selenium
PyYAML
pika
nose
psutil
pymysql
